requirements.txt
python-telegram-bot
psycopg2-binary
pandas
openpyxl
python-dotenv